+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/calculation/CraftingState.java
@@ -3,21 +3,18 @@ package com.refinedmods.refinedstorage.api.autocrafting.calculation;
 import com.refinedmods.refinedstorage.api.autocrafting.Pattern;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
+import com.refinedmods.refinedstorage.api.resource.list.LazyCopyMutableResourceList;
 import com.refinedmods.refinedstorage.api.resource.list.MutableResourceList;
 import com.refinedmods.refinedstorage.api.resource.list.MutableResourceListImpl;
 import com.refinedmods.refinedstorage.api.storage.root.RootStorage;
 
-import java.util.Comparator;
-
 class CraftingState {
     private final MutableResourceList storage;
     private final MutableResourceList internalStorage;
-    private final Comparator<ResourceKey> sorter;
 
     private CraftingState(final MutableResourceList storage, final MutableResourceList internalStorage) {
         this.storage = storage;
         this.internalStorage = internalStorage;
-        this.sorter = createSorter(storage).thenComparing(createSorter(internalStorage));
     }
 
     void extractFromInternalStorage(final ResourceKey resource, final long amount) {
@@ -47,32 +44,31 @@ class CraftingState {
     static CraftingState of(final RootStorage rootStorage) {
         final MutableResourceListImpl storage = MutableResourceListImpl.create();
         rootStorage.getAll().forEach(storage::add);
-        return new CraftingState(storage, MutableResourceListImpl.create());
-    }
-
-    Comparator<ResourceKey> getSorter() {
-        return sorter;
-    }
-
-    private static Comparator<ResourceKey> createSorter(final MutableResourceList list) {
-        return (a, b) -> {
-            final long ar = list.get(a);
-            final long br = list.get(b);
-            return (int) br - (int) ar;
-        };
+        return new CraftingState(LazyCopyMutableResourceList.create(storage), MutableResourceListImpl.create());
     }
 
     ResourceState getResource(final ResourceKey resource) {
         return new ResourceState(resource, storage.get(resource), internalStorage.get(resource));
     }
 
-    record ResourceState(ResourceKey resource, long inStorage, long inInternalStorage) {
+    record ResourceState(ResourceKey resource, long inStorage, long inInternalStorage)
+        implements Comparable<ResourceState> {
         boolean isInStorage() {
             return inStorage > 0;
         }
 
         boolean isInInternalStorage() {
             return inInternalStorage > 0;
+        }
+
+        @Override
+        public int compareTo(final ResourceState o) {
+            // o first, we want greater values to come first
+            final int storage = Long.compare(o.inStorage, inStorage);
+            if (storage == 0) {
+                return Long.compare(o.inInternalStorage, inInternalStorage);
+            }
+            return storage;
         }
     }
 }