+++ refinedstorage-resource-api/src/test/java/com/refinedmods/refinedstorage/api/resource/list/ProxyResourceListTest.java
@@ -1,9 +1,15 @@
 package com.refinedmods.refinedstorage.api.resource.list;
 
+import com.refinedmods.refinedstorage.api.resource.TestResource;
+
 class ProxyResourceListTest extends AbstractMutableResourceListTest {
     @Override
-    protected MutableResourceList createList() {
-        return new AbstractProxyMutableResourceList(MutableResourceListImpl.create()) {
+    protected MutableResourceList createList(final TestResource[] resources, final long amount) {
+        final MutableResourceListImpl delegate = MutableResourceListImpl.create();
+        for (final TestResource resource : resources) {
+            delegate.add(resource, amount);
+        }
+        return new AbstractProxyMutableResourceList(delegate) {
         };
     }
 }